refactor: 로또 등수를 체크하는 checkLottoRank()를 LottoTicket 클래스로 이동
